Building with separate peripheral source files
