+++ Core/Inc/ltdc.h
@@ -1,0 +1,54 @@
+/* USER CODE BEGIN Header */
+/**
+  ******************************************************************************
+  * @file    ltdc.h
+  * @brief   This file contains all the function prototypes for
+  *          the ltdc.c file
+  ******************************************************************************
+  * @attention
+  *
+  * Copyright (c) 2024 STMicroelectronics.
+  * All rights reserved.
+  *
+  * This software is licensed under terms that can be found in the LICENSE file
+  * in the root directory of this software component.
+  * If no LICENSE file comes with this software, it is provided AS-IS.
+  *
+  ******************************************************************************
+  */
+/* USER CODE END Header */
+/* Define to prevent recursive inclusion -------------------------------------*/
+#ifndef __LTDC_H__
+#define __LTDC_H__
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/* Includes ------------------------------------------------------------------*/
+#include "main.h"
+
+/* USER CODE BEGIN Includes */
+#include "dsihost.h"
+#include "../Components/nt35510/nt35510.h"
+/* USER CODE END Includes */
+
+extern LTDC_HandleTypeDef hltdc;
+
+/* USER CODE BEGIN Private defines */
+#define LCD_ORIENTATION_LANDSCAPE 0x01
+
+/* USER CODE END Private defines */
+
+void MX_LTDC_Init(void);
+
+/* USER CODE BEGIN Prototypes */
+
+/* USER CODE END Prototypes */
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* __LTDC_H__ */
+
