+++ STM32F469I-DISCO-REV-AU1.ioc
@@ -413,7 +413,7 @@ ProjectManager.AskForMigrate=true
 ProjectManager.BackupPrevious=false
 ProjectManager.CompilerOptimize=6
 ProjectManager.ComputerToolchain=false
-ProjectManager.CoupleFile=false
+ProjectManager.CoupleFile=true
 ProjectManager.CustomerFirmwarePackage=
 ProjectManager.DefaultFWLocation=true
 ProjectManager.DeletePrevious=true
@@ -439,7 +439,7 @@ ProjectManager.ToolChainLocation=
 ProjectManager.UAScriptAfterPath=
 ProjectManager.UAScriptBeforePath=
 ProjectManager.UnderRoot=false
-ProjectManager.functionlistsort=1-MX_GPIO_Init-GPIO-false-HAL-true,2-SystemClock_Config-RCC-false-HAL-false,3-MX_CRC_Init-CRC-false-HAL-true,4-MX_DMA2D_Init-DMA2D-false-HAL-true,5-MX_DSIHOST_DSI_Init-DSIHOST-false-HAL-true,6-MX_FMC_Init-FMC-false-HAL-true,7-MX_LTDC_Init-LTDC-false-HAL-true,8-MX_QUADSPI_Init-QUADSPI-false-HAL-true,9-MX_I2C1_Init-I2C1-false-HAL-true,10-MX_CAN2_Init-CAN2-false-HAL-true,11-MX_USART6_UART_Init-USART6-false-HAL-true,12-MX_SDIO_SD_Init-SDIO-false-HAL-true,13-MX_TouchGFX_Init-STMicroelectronics.X-CUBE-TOUCHGFX.4.22.1-false-HAL-false,14-MX_TouchGFX_Process-STMicroelectronics.X-CUBE-TOUCHGFX.4.22.1-false-HAL-false
+ProjectManager.functionlistsort=1-MX_GPIO_Init-GPIO-false-HAL-true,2-SystemClock_Config-RCC-false-HAL-false,3-MX_CRC_Init-CRC-false-HAL-true,4-MX_DMA2D_Init-DMA2D-false-HAL-true,5-MX_DSIHOST_DSI_Init-DSIHOST-false-HAL-true,6-MX_FMC_Init-FMC-false-HAL-true,7-MX_LTDC_Init-LTDC-false-HAL-true,8-MX_QUADSPI_Init-QUADSPI-false-HAL-true,9-MX_I2C1_Init-I2C1-false-HAL-true,10-MX_CAN2_Init-CAN2-false-HAL-true,11-MX_USART6_UART_Init-USART6-false-HAL-true,12-MX_SDIO_SD_Init-SDIO-false-HAL-true,13-MX_FATFS_Init-FATFS-false-HAL-false,14-MX_TouchGFX_Init-STMicroelectronics.X-CUBE-TOUCHGFX.4.22.1-false-HAL-false,15-MX_TouchGFX_Process-STMicroelectronics.X-CUBE-TOUCHGFX.4.22.1-false-HAL-false
 QUADSPI.ChipSelectHighTime=QSPI_CS_HIGH_TIME_5_CYCLE
 QUADSPI.ClockPrescaler=1
 QUADSPI.FlashSize=27
